NPUW: Disable AVX2 code with ENABLE_AVX2=OFF  (#26890)
### Details:
 - Disable AVX2 code with ENABLE_AVX2=OFF

### Tickets:
 - E-141645
